Merge MySQL 9.7 - Xtrabackup 9.7.1-rc1

+++ storage/innobase/xtrabackup/test/t/simulate_different_version.sh
@@ -51,7 +51,49 @@ fi
 
 rm $topdir/pxb.log
 
-vlog "case#5 backup should succesful with no-server-version-check and 9.1.99 version"
+vlog "case#5 backup should fail with 9.6 version (older 9.x, not 9.7.x)"
+
+run_cmd_expect_failure $XB_BIN $XB_ARGS --backup --debug=d,simulate_96_version --target-dir=$topdir/backup 2>&1 | tee $topdir/pxb.log
+
+if ! grep -q "Unsupported server version: '9.6.99'" $topdir/pxb.log
+then
+  die "xtrabackup did not show 'Unsupported server version' error for 9.6.99"
+fi
+
+if ! grep -q 'only supports backing up and restoring MySQL and Percona Servers of version 9.7.x' $topdir/pxb.log
+then
+  die "xtrabackup did not show 9.7.x-only support message for 9.6.99"
+fi
+
+if ! grep -q 'Please use Percona XtraBackup 9.6 for this database.' $topdir/pxb.log
+then
+  die "xtrabackup did not suggest Percona XtraBackup 9.6 for a 9.6 server"
+fi
+
+rm $topdir/pxb.log
+
+vlog "case#6 backup should fail with 9.8 version (newer 9.x, not 9.7.x)"
+
+run_cmd_expect_failure $XB_BIN $XB_ARGS --backup --debug=d,simulate_98_version --target-dir=$topdir/backup 2>&1 | tee $topdir/pxb.log
+
+if ! grep -q "Unsupported server version: '9.8.0'" $topdir/pxb.log
+then
+  die "xtrabackup did not show 'Unsupported server version' error for 9.8.0"
+fi
+
+if ! grep -q 'only supports backing up and restoring MySQL and Percona Servers of version 9.7.x' $topdir/pxb.log
+then
+  die "xtrabackup did not show 9.7.x-only support message for 9.8.0"
+fi
+
+if ! grep -q 'Please use Percona XtraBackup 9.8 for this database.' $topdir/pxb.log
+then
+  die "xtrabackup did not suggest Percona XtraBackup 9.8 for a 9.8 server"
+fi
+
+rm $topdir/pxb.log
+
+vlog "case#7 backup should succeed on a simulated 9.7.99 server (highest 9.7.x patch)"
 xtrabackup --backup --debug=d,simulate_higher_version --no-server-version-check --target-dir=$topdir/backup
 xtrabackup --prepare --debug=d,simulate_higher_version --no-server-version-check --target-dir=$topdir/backup
 rm -r $topdir/backup
